Release v0.0.7 (#115)
* fix: npm publish action (#86)
* chore: adds publish web to publish action
* feat: fraxBP vaults (#71)
* feat: test for update assetsPerShare after removal of underlyingVault
* feat: ConvexFrax based vaults
* chore: Refactor Curve lib tests (#74)
* chore: using Hardhat network helpers for setBalance, setStorageAt and stopImpersonateAccount (#82)
* fix: missing awaits in savePlus fork tests
* refact: reduce gas consumption while donating tokens (#87)
* chore: updated docs for second audit
* chore: adds deployment address (#93)
* fix: removeVault() will block because of incorrect vaults length (#96)
* fix: modified require to use total vaults length instead of active vaults
* feat: test for nth vault removal
* fix: outdated assets amount in _redeem and _mint for PeriodicAllocationAbstractVault (#99)
* feat: recalculate assets amount of successful assetsPerShare update
* feat: tests for recalculating assetsPerShare for redeem and mint
* fix: Meta vaults cannot re-add underlying vaults after removal (#100)
* feat: removed allowance in remove vault
* fix: test for remove and re-add vault
* fix: liq-donate-tokens task with multiple rewards (#101)
* fix: liq-settle-swap task (#103)
* fix: removes death code from LiquidatorVaults (#104)
* fix: RoundUp for `mint`, `previewMint`, `withdraw` and `previewWithdraw` (#102)
* feat: contract changes for roundUp
* feat: LightBasicVault roundUp
* feat: roundUp unit-tests
* feat: 3Crv based vault changes
* feat: FraxBp based vault changes
* feat: roundUp tests and scaling bug fix
* fix: reset network before running test
* fix: [Macro] Medium and lower criticality fixes (#106)
* feat: [M-4] added check for performance fees too high during initialization
* feat: [L-2] cannot remove cached vault
* feat: [Q-1] removed zero-address check for code consistency
* feat: [Q-2] removed unused imports
* feat: [Q-3] corrected comment
* feat: [G-1] No check if performanceFee is 0
* feat: [G-3] Unnecessary control variable initialization of for loops And removed zero address check vaultManagerRole
* fix: fork-tests for sameAssetsUnderlyingVaults behavior for removeVault
* feat: [G-2] uint24 for performanceFee
* fix: uint24 performanceFee
* fix: unit-test errors
* fix: removed toNumber()
* chore: toNumber fix
* fix: minor fixes
* feat: updated diagrams
* feat: get historical token balances using token-balance HH task (#107)
* fix: liquidator retry approvals (#110)
* fix: liquidator retry approvals by pre-approving the async swapper to transfer reward tokens
* fix: typos in README (#112)

+++ contracts/vault/fee/PerfFeeBasicVault.sol
@@ -33,7 +33,7 @@ contract PerfFeeBasicVault is Initializable, PerfFeeAbstractVault {
         string calldata _symbolArg,
         address _vaultManager,
         address feeReceiver,
-        uint32 _performanceFee
+        uint24 _performanceFee
     ) external initializer {
         // Set the vault's decimals to the same as the reference asset.
         uint8 decimals = InitializableToken(address(_asset)).decimals();
